requirements.txt
tensorflow==2.15.0
numpy
matplotlib
scikit-learn
seaborn
pandas
jupyter
Pillow
opencv-python

# pip install -r requirements.txt
